Vendor_email_logger_agent/requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
supabase>=1.0.3
python-dotenv>=1.0.0
aiohttp>=3.8.4
asyncio==3.4.3
python-dateutil>=2.8.2
openai>=1.3.0
numpy>=1.24.3
pandas>=2.0.3
tiktoken>=0.5.1 